Restructured files to resemble a real world website (/login, /register, and /logout would not come after /users in real life and thus they are in the server now). Ditto for the protected page. Good reminder of what to import where. Also created a new account and tested all resources to make sure they still behave as expected.
